Add files via upload
To develop a system for the detection of Covid-19 samples from healthy and Covid-19 positive cases using X-ray images.
To develop a quick, easy and cost-efficient procedure for detection. We propose a deep learning CNN architecture for identifying Covid-19 positive patients using chest X-ray images.
We collected dataset from Kaggle. The dataset id divided for training and testing. 80% is used for training and 20% of dataset for testing. The dataset is comprised of 5 classes : Normal, Type1, Type2, Type3 and Type4. 
Chest X-ray images are given as datasets to the trained CNN model. After testing the X-ray image, the CNN returns the result.
Covid-19 positive or normal case.